+++ lib/signin/signin.dart
@@ -1,0 +1,3 @@
+export 'view/signin_page.dart';
+export 'widgets/google_sign_in_button.dart';
+export 'widgets/user_info.dart';
